+++ src/interfaces_graficas/AltaCliente.java
@@ -615,7 +615,7 @@ public class AltaCliente extends JPanel {
 					this.textFieldNombre.getText(), 
 					this.textFieldLocalidad.getText(),
 					this.textFieldApellido.getText(),
-					Integer.parseInt(this.textFieldTelefono.getText()),
+					Long.parseLong(this.textFieldTelefono.getText()),
 					this.textFieldBarrio.getText(),
 					Integer.parseInt(this.textFieldMonto.getText()));
